AI_Agent_for_AutoMobiles/requirements.txt
langchain
langchain_community==0.2.7
streamlit
crewai
crew_tools==0.4.8
python-dotenv 
